refactor links to its class cmpt
